Move Gaze into Utility namespace for consistency.
Also renames function to camelCase for consistency.
